requirements.txt
asgiref==3.8.1
boto3==1.35.60
botocore==1.35.60
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
websockets==14.1
pytz